+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Unreleased
 ----------
+- Changed "front" and "back" semantics to more closely resemble those of
+  `VecDeque`
 - Removed `Deref` and `DerefMut` impls from `RingBuf`
 - Removed `RingBuf::front_idx` and `RingBuf::back_idx` methods
 - Bumped minimum supported Rust version to `1.58`